Move list, edit and details related code to their own directories
