+++ web/html/src/manager/maintenance/details/calendar-details.js
