+++ web/html/src/manager/maintenance/details/maintenance-windows-details.js
