+++ web/html/src/manager/maintenance/details/schedule-details.js
