+++ web/html/src/manager/maintenance/edit/calendar-edit.js
