+++ web/html/src/manager/maintenance/edit/schedule-edit.js
