+++ web/html/src/manager/maintenance/list/calendar-list.js
