+++ web/html/src/manager/maintenance/list/maintenance-windows-list.js
