+++ web/html/src/manager/maintenance/list/schedule-list.js
