+++ web/html/src/manager/maintenance/maintenance-windows.js
@@ -5,9 +5,9 @@ const React = require("react");
 const ReactDOM = require("react-dom");
 const Messages = require("components/messages").Messages;
 const Network = require("utils/network");
-const {MaintenanceWindowsDetails} = require("./maintenance-windows-details");
-const {MaintenanceWindowsList} = require("./maintenance-windows-list");
-const {MaintenanceWindowsEdit} =  require("./maintenance-windows-edit");
+const {MaintenanceWindowsDetails} = require("./details/maintenance-windows-details");
+const {MaintenanceWindowsList} = require("./list/maintenance-windows-list");
+const {MaintenanceWindowsEdit} =  require("./edit/maintenance-windows-edit");
 const MessagesUtils = require("components/messages").Utils;
 const SpaRenderer  = require("core/spa/spa-renderer").default;
 const MaintenanceWindowsApi = require("./maintenance-windows-api");
